LayoutTests/webexposed/counter-style-image-symbols-not-exposed-expected.txt


FAIL @counter-style image symbols are not exposed assert_equals: Should've been able to parse @counter-style. expected 1 but got 0
